fix: fixes in the generate template functionality

+++ utils/generator.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const path = require('path');
 
 const { fileExists, writeToPath } = require('./file.utils');
@@ -14,29 +15,35 @@ function createFiles(name) {
 
   const pathName = path.join(currentWorkingDirectory, '/src/components');
 
-  const componentPath = path.join(pathName, _name);
+  const doesSrcFolderExists = fs.existsSync(pathName);
+  if (!doesSrcFolderExists) {
+    console.log('⚠️  Could not find /src/components directory in the project. ');
+    console.log('⚠️  Are you running the command in a react-js project? ');
+  } else {
+    const componentPath = path.join(pathName, _name);
 
-  const files = {
-    component: `${_name}.jsx`,
-    test: `${_name}.test.js`,
-    style: `${_name}.css`,
-    storybook: `${_name}.stories.jsx`,
-  };
+    const files = {
+      component: `${_name}.jsx`,
+      test: `${_name}.test.js`,
+      style: `${_name}.css`,
+      storybook: `${_name}.stories.jsx`,
+    };
 
-  if (_name !== 'components') {
-    const writeFile = writeToPath(pathName);
-    const toFileMissingBool = (file) => !fileExists(componentPath)(file);
-    const checkAllMissing = (acc, cur) => acc && cur;
+    if (_name !== 'components') {
+      const writeFile = writeToPath(pathName);
+      const toFileMissingBool = (file) => !fileExists(componentPath)(file);
+      const checkAllMissing = (acc, cur) => acc && cur;
 
-    const noneExist = Object.values(files).map(toFileMissingBool).reduce(checkAllMissing);
+      const noneExist = Object.values(files).map(toFileMissingBool).reduce(checkAllMissing);
 
-    if (noneExist) {
-      Object.entries(files).forEach(([type, fileName]) => {
-        writeFile(fileName, templates[type](_name));
-      });
-      console.log('🚀 Happy Hacking ! 🚀');
-    } else {
-      console.log(`⚠️Component with name ${_name} already exists!`);
+      if (noneExist) {
+        Object.entries(files).forEach(([type, fileName]) => {
+          writeFile(fileName, templates[type](_name));
+        });
+        console.log('🚀 Happy Hacking ! 🚀');
+      } else {
+        console.log(`⚠️ Component with name ${_name} already exists!`);
+      }
     }
   }
 }
